requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.0.1
cachetools==6.0.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
hiredis==3.2.1
idna==3.10
mypy==1.16.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
passlib==1.7.4
pathspec==0.12.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlalchemy-stubs==0.4
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
